+++ packages/terra-navigation-prompt/src/NavigationPromptCheckpoint.jsx
@@ -133,10 +133,13 @@ class NavigationPromptCheckpoint extends React.Component {
       this.setState({
         notificationDialogProps: {
           title: showDialogOptions.title,
-          message: showDialogOptions.message,
+          startMessage: showDialogOptions.startMessage,
+          content: showDialogOptions.content,
+          endMessage: showDialogOptions.endMessage,
           acceptButtonText: showDialogOptions.acceptButtonText,
           rejectButtonText: showDialogOptions.rejectButtonText,
           emphasizedAction: showDialogOptions.emphasizedAction,
+          buttonOrder: showDialogOptions.buttonOrder,
           onAccept: resolve,
           onReject: reject,
         },
@@ -146,17 +149,14 @@ class NavigationPromptCheckpoint extends React.Component {
 
   renderNotificationDialog() {
     const {
-      title, message, acceptButtonText, rejectButtonText, emphasizedAction, onAccept, onReject,
+      title, startMessage, endMessage, content, acceptButtonText, rejectButtonText, emphasizedAction, buttonOrder, onAccept, onReject,
     } = this.state.notificationDialogProps;
-
-    const acceptActionIsEmphasized = !emphasizedAction || emphasizedAction === 'accept';
 
     const acceptButton = {
       text: acceptButtonText,
       onClick: () => {
         this.setState({ notificationDialogProps: undefined }, onAccept);
       },
-      isEmphasized: (acceptActionIsEmphasized),
     };
 
     const rejectButton = {
@@ -164,16 +164,19 @@ class NavigationPromptCheckpoint extends React.Component {
       onClick: () => {
         this.setState({ notificationDialogProps: undefined }, onReject);
       },
-      isEmphasized: (!acceptActionIsEmphasized),
     };
 
     return (
       <NotificationDialog
         isOpen
         title={title}
-        startMessage={message}
-        acceptAction={acceptActionIsEmphasized ? acceptButton : rejectButton}
-        rejectAction={acceptActionIsEmphasized ? rejectButton : acceptButton}
+        startMessage={startMessage}
+        endMessage={endMessage}
+        content={content}
+        acceptAction={acceptButton}
+        rejectAction={rejectButton}
+        buttonOrder={buttonOrder}
+        emphasizedAction={emphasizedAction}
         variant="warning"
       />
     );