+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/CompleteNotificationDialog.jsx
@@ -42,6 +42,7 @@ class CompleteNotificationDialog extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />