+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithFocus.jsx
@@ -59,6 +59,7 @@ class NotificationDialogWithFocus extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />