+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithLongText.jsx
@@ -121,6 +121,7 @@ class NotificationDialogWithLongText extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />