+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/CompleteNotificationDialog.test.jsx
@@ -42,6 +42,7 @@ class CompleteNotificationDialog extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />