+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogMinimalProps.test.jsx
@@ -31,6 +31,7 @@ class NotificationDialogMinimalProps extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />