+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogNoMessage.test.jsx
@@ -33,6 +33,7 @@ class NotificationDialogNoMessage extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />