+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogVariant.test.jsx
@@ -49,6 +49,7 @@ class NotificationDialogVariant extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          buttonOrder="acceptFirst"
           emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />