+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/ReversedActionNotificationDialog.test.jsx
@@ -42,7 +42,7 @@ class ReversedActionNotificationDialog extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
-          isReversed
+          buttonOrder="rejectFirst"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>